fix : ai api와 ai db 관련 서비스 분리
